requirements.txt
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
faiss-cpu==1.8.0
langchainhub==0.1.15
python-dotenv==1.0.1
langchain-text-splitters==0.0.1
discord